requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
motor==3.6.0
pydantic==2.9.0
python-dotenv==1.0.1
email-validator==2.2.0
google-generativeai

python-jose[cryptography]
passlib[bcrypt]
python-multipart